backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
reportlab==4.0.7
email-validator==2.1.0.post1
numpy==1.26.4
scipy==1.11.4
opencv-python-headless==4.10.0.84
torch==2.3.1
openai==1.51.2
python-dotenv==1.0.1
